pyproject.toml
[project]
name = "sample-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "starlette-context>=0.4.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "ipython>=9.4.0",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.2",
    "types-redis>=4.6.0.20241004",
]

[tool.ruff]
indent-width = 4
line-length = 120
include = ["src/**/*.py"]
exclude = ["src/tests/**", "src/migrations/**"]
cache-dir = ".ruff_cache"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "F821"]
fixable = ["E", "F", "W", "I"]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_redefinition = false
